svn merge -r 42811:42904 http://svn.civicrm.org/civicrm/branches/v4.2 -x -w --ignore-ancestry

+++ CRM/Core/Extensions.php
@@ -123,6 +123,7 @@ class CRM_Core_Extensions {
         }
       }
       else {
+        if (CRM_Core_Permission::check('administer CiviCRM') && $this->isDownloadEnabled()) {
         $civicrmDestination = urlencode(CRM_Utils_System::url('civicrm/admin/extensions', 'reset=1'));
         $url = CRM_Utils_System::url('civicrm/admin/setting/path', "reset=1&civicrmDestination=${civicrmDestination}");
         CRM_Core_Session::setStatus(ts('Your extensions directory: %1 is not web server writable. Please go to the <a href="%2">path setting page</a> and correct it.<br/>',
@@ -131,10 +132,11 @@ class CRM_Core_Extensions {
               2 => $url,
             )
           ), ts('Directory Unwritable'), 'alert');
+        }
         $this->_extDir = NULL;
       }
 
-      if (!class_exists('ZipArchive')) {
+      if (!class_exists('ZipArchive') && CRM_Core_Permission::check('administer CiviCRM') && $this->isDownloadEnabled()) {
         // everyone else is dumping messages wily-nily, why can't I?
         CRM_Core_Session::setStatus(
           ts('You will not be able to install extensions at this time because your installation of PHP does not support ZIP archives. Please ask your system administrator to install the standard PHP-ZIP extension.'),
@@ -142,7 +144,7 @@ class CRM_Core_Extensions {
         );
     }
 
-      if (empty($config->extensionsURL)) {
+      if (empty($config->extensionsURL) && CRM_Core_Permission::check('administer CiviCRM')) {
         $civicrmDestination = urlencode(CRM_Utils_System::url('civicrm/admin/extensions', 'reset=1'));
         $url = CRM_Utils_System::url('civicrm/admin/setting/url', "reset=1&civicrmDestination=${civicrmDestination}");
         CRM_Core_Session::setStatus(ts('Your Extensions Directory (%1) does not have a matching Extensions Resource URL. Please go to the <a href="%2">URL setting page</a> and correct it.<br/>',
@@ -816,6 +818,15 @@ class CRM_Core_Extensions {
    */
   public function isEnabled() {
       return $this->enabled;
+}
+
+  /**
+   * Determine whether the system allows downloading new extensions
+   *
+   * @return bool
+   */
+  public function isDownloadEnabled() {
+    return (FALSE !== $this->getRepositoryUrl());
 }
 }
